Refactor FeaturesRegistryService into modular package
Break down 585-line monolithic service into focused components:
- ProviderManager: Provider availability, validation, and aliases
- NLPCapabilityChecker: fuzzywuzzy and spaCy detection
- FeaturesRegistryService: Facade coordinating specialized components

Maintains backwards compatibility via re-exports from original module path.
